bi-agent/docker/company_agent/requirements.txt
sqlalchemy
psycopg2-binary
pymysql
pandas
requests
pyyaml
schedule
